sc2: Split usecase unit tests from item filtering tests
